refactor: update navigation structure
