+++ composeApp/src/commonMain/kotlin/com/jefisu/trackizer/navigation/Destination.kt
@@ -8,7 +8,7 @@ sealed interface Destination {
     data object AuthGraph : Destination
 
     @Serializable
-    data object AuthenticatedGraph : Destination
+    data object LoggedGraph : Destination
 
     @Serializable
     data object WelcomeScreen : Destination
