+++ composeApp/src/commonMain/kotlin/com/jefisu/trackizer/navigation/NavExtension.kt
@@ -1,12 +1,16 @@
 package com.jefisu.trackizer.navigation
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.remember
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination.Companion.hasRoute
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.jefisu.trackizer.core.ui.LocalAnimatedContentScope
 import org.koin.compose.currentKoinScope
 import org.koin.compose.viewmodel.koinViewModel
 import org.koin.core.scope.Scope
@@ -34,5 +38,15 @@ inline fun <reified T : Destination> NavController.navigateSingleTopTo(destinati
     navigate(destination) {
         popUpTo<T>()
         launchSingleTop = true
+    }
+}
+
+inline fun <reified T : Destination> NavGraphBuilder.animatedScreen(
+    noinline content: @Composable NavBackStackEntry.() -> Unit,
+) {
+    composable<T> { backStack ->
+        CompositionLocalProvider(LocalAnimatedContentScope provides this) {
+            content(backStack)
+        }
     }
 }
