+++ composeApp/src/commonMain/kotlin/com/jefisu/trackizer/navigation/NavGraph.kt
@@ -7,45 +7,22 @@ import androidx.compose.animation.SharedTransitionLayout
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.remember
-import androidx.navigation.NavBackStackEntry
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.navigation
-import com.jefisu.trackizer.auth.di.rememberAuthScope
-import com.jefisu.trackizer.auth.presentation.login.LoginScreenRoot
-import com.jefisu.trackizer.auth.presentation.register.RegisterScreenRoot
-import com.jefisu.trackizer.auth.presentation.thirdpartyauth.ThirdPartyAuthRoot
-import com.jefisu.trackizer.core.ui.LocalAnimatedContentScope
 import com.jefisu.trackizer.core.ui.LocalSharedTransitionScope
-import com.jefisu.trackizer.home.HomeScreenRoot
-import com.jefisu.trackizer.welcome.WelcomeScreenRoot
+import com.jefisu.trackizer.navigation.graph.authGraph
+import com.jefisu.trackizer.navigation.graph.authNavAnimConfig
+import com.jefisu.trackizer.navigation.graph.loggedGraph
+import com.jefisu.trackizer.navigation.graph.loggedNavAnimConfig
 
 @Composable
 fun NavGraph(startDestination: Destination, navController: NavHostController) {
     val navAnimationConfig = remember {
-        listOf(
-            Destination.WelcomeScreen to listOf(
-                AnimationTarget(Destination.LoginScreen, AnimationType.BOTTOM_TO_TOP),
-                AnimationTarget(Destination.ThirdPartyAuthScreen, AnimationType.BOTTOM_TO_TOP),
-                AnimationTarget(Destination.RegisterScreen, AnimationType.BOTTOM_TO_TOP),
-            ),
-            Destination.ThirdPartyAuthScreen to listOf(
-                AnimationTarget(Destination.RegisterScreen, AnimationType.RIGHT_TO_LEFT),
-                AnimationTarget(Destination.HomeScreen, AnimationType.BOTTOM_TO_TOP),
-            ),
-            Destination.LoginScreen to listOf(
-                AnimationTarget(Destination.ThirdPartyAuthScreen, AnimationType.TOP_TO_BOTTOM),
-                AnimationTarget(Destination.HomeScreen, AnimationType.BOTTOM_TO_TOP),
-            ),
-            Destination.RegisterScreen to listOf(
-                AnimationTarget(Destination.LoginScreen, AnimationType.LEFT_TO_RIGHT),
-                AnimationTarget(Destination.ThirdPartyAuthScreen, AnimationType.TOP_TO_BOTTOM),
-            ),
-        ).map { (currentDestination, targets) ->
-            AnimationConfig(currentDestination = currentDestination, targets = targets)
+        buildList {
+            addAll(authNavAnimConfig)
+            addAll(loggedNavAnimConfig)
+        }.map { (currentDestination, targets) ->
+            AnimationConfig(currentDestination, targets)
         }
     }
     val navAnimation = remember { NavAnimation(navAnimationConfig) }
@@ -69,73 +46,9 @@ fun NavGraph(startDestination: Destination, navController: NavHostController) {
                     },
                 ) {
                     authGraph(navController)
-                    authenticatedGraph()
+                    loggedGraph()
                 }
             }
-        }
-    }
-}
-
-private fun NavGraphBuilder.authGraph(navController: NavController) {
-    fun navigate(destination: Destination) {
-        navController.navigateSingleTopTo<Destination.WelcomeScreen>(destination)
-    }
-
-    navigation<Destination.AuthGraph>(
-        startDestination = Destination.WelcomeScreen,
-    ) {
-        animatedScreen<Destination.WelcomeScreen> {
-            WelcomeScreenRoot(
-                onNavigateToSignInScreen = {
-                    navigate(Destination.LoginScreen)
-                },
-                onNavigateToSignUpScreen = {
-                    navigate(Destination.ThirdPartyAuthScreen)
-                },
-            )
-        }
-        animatedScreen<Destination.ThirdPartyAuthScreen> {
-            ThirdPartyAuthRoot(
-                onNavigateToRegisterScreen = {
-                    navigate(Destination.RegisterScreen)
-                },
-            )
-        }
-        animatedScreen<Destination.LoginScreen> {
-            LoginScreenRoot(
-                viewModel = sharedViewModel(navController, rememberAuthScope()),
-                onNavigateToRegisterScreen = {
-                    navigate(Destination.ThirdPartyAuthScreen)
-                },
-            )
-        }
-        animatedScreen<Destination.RegisterScreen> {
-            RegisterScreenRoot(
-                viewModel = sharedViewModel(navController, rememberAuthScope()),
-                onNavigateToLoginScreen = {
-                    navigate(Destination.LoginScreen)
-                },
-            )
-        }
-    }
-}
-
-private fun NavGraphBuilder.authenticatedGraph() {
-    navigation<Destination.AuthenticatedGraph>(
-        startDestination = Destination.HomeScreen,
-    ) {
-        animatedScreen<Destination.HomeScreen> {
-            HomeScreenRoot()
-        }
-    }
-}
-
-private inline fun <reified T : Destination> NavGraphBuilder.animatedScreen(
-    noinline content: @Composable NavBackStackEntry.() -> Unit,
-) {
-    composable<T> { backStack ->
-        CompositionLocalProvider(LocalAnimatedContentScope provides this) {
-            content(backStack)
         }
     }
 }
